requirements-fixed.txt
# Updated Requirements for GoFile Uploader Bot

pyTelegramBotAPI==4.29.1
aiohttp==3.9.5
aiofiles==23.2.0
motor==3.4.0
pymongo==4.7.3
requests==2.31.0
python-dotenv==1.0.1
python-dateutil==2.9.0